Tooling updates. Use file set for headers. (#19)
### Why 
C++ is a moving target.

Updated to clang-19 since clang-20 in c++23 mode has issues with the
version of fmt that is available on vcpkg.
Updated to gcc-14 since gcc-15 is not easily available in the Ubuntu
24.04 apt repositories.
Still kept gcc-11, since that is what vcpkg will use.

I will update required CI actions, if approved.

### What was changed
- **Use IWYU pragmas over NOLINTs.**
- **Update toolchains. Use clang-19 and gcc-14 in dev and CI**
- **Use FILE_SET HEADERS. Update cmake-format configuration.**
- **Update vcpkg baseline.**
- **Version bump**